Allow decoration of serialization context.
